Update package.json and application routes***
This commit updates the "main" file in the package.json of the aruba-connection package to "app.js" instead of "aruba-ws-test-endpoint.js". It also modifies the application routes in the front-end/src/application.route.tsx file to redirect to "/telemetry-websocket" for both the root path and any other unmatched paths. Additionally, it adds a new function "encodeHexStringToBase64" in the front-end/src/pages/connector/index.tsx file to encode hexadecimal strings to base64. Finally, it modifies the "onFinish" function to use the "encodeHexStringToBase64" function to encode certain values before making a POST request.